feat(cli): add project management commands for init, apply, check, edit, info, run, and require
feat(cli): implement rename command for environment variables with pattern matching

feat(cli): add replace command for updating environment variable values and find/replace functionality

feat(cli): introduce snapshot management commands for creating, listing, showing, restoring, deleting, and diffing snapshots

feat(cli): implement watch command for file synchronization with environment variables

fix(core): enhance EnvWatcher with change logging, pattern matching, and file handling capabilities

test(core): add comprehensive tests for EnvWatcher, including file watching, change logging, and synchronization scenarios
